move BasePaparazziTest to a :ui:base testFixture
