+++ app/src/testDebug/kotlin/org/cru/godtools/ui/dashboard/home/AllFavoritesLayoutPaparazziTest.kt
@@ -19,7 +19,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
-import org.cru.godtools.ui.BasePaparazziTest
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.ui.dashboard.home.AllFavoritesScreen.UiState
 import org.cru.godtools.ui.tools.ToolCardStateTestData
 import org.junit.runner.RunWith
