+++ app/src/testDebug/kotlin/org/cru/godtools/ui/dashboard/tools/ToolFiltersPaparazziTest.kt
@@ -11,8 +11,8 @@ import java.util.Locale
 import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlinx.collections.immutable.persistentListOf
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.model.Language
-import org.cru.godtools.ui.BasePaparazziTest
 import org.cru.godtools.ui.dashboard.filters.FilterMenu.UiState
 import org.junit.runner.RunWith
