+++ app/src/testDebug/kotlin/org/cru/godtools/ui/languages/downloadable/LanguageDownloadStatusIndicatorPaparazziTest.kt
@@ -6,7 +6,7 @@ import com.android.resources.NightMode
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
 import kotlin.test.Test
-import org.cru.godtools.ui.BasePaparazziTest
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.junit.runner.RunWith
 
 @RunWith(TestParameterInjector::class)
