+++ app/src/testDebug/kotlin/org/cru/godtools/ui/tooldetails/ToolDetailsLayoutPaparazziTest.kt
@@ -20,11 +20,11 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.downloadmanager.DownloadProgress
 import org.cru.godtools.model.Language
 import org.cru.godtools.model.randomTool
 import org.cru.godtools.model.randomTranslation
-import org.cru.godtools.ui.BasePaparazziTest
 import org.cru.godtools.ui.drawer.DrawerMenuScreenStateTestData
 import org.cru.godtools.ui.tooldetails.ToolDetailsScreen.State
 import org.cru.godtools.ui.tools.ToolCard
