+++ app/src/testDebug/kotlin/org/cru/godtools/ui/tools/SquareToolCardPaparazziTest.kt
@@ -17,8 +17,8 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.downloadmanager.DownloadProgress
-import org.cru.godtools.ui.BasePaparazziTest
 import org.junit.runner.RunWith
 
 @RunWith(TestParameterInjector::class)
