+++ app/src/testDebug/kotlin/org/cru/godtools/ui/tools/ToolCardPaparazziTest.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
-import org.cru.godtools.ui.BasePaparazziTest
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.junit.runner.RunWith
 
 @RunWith(TestParameterInjector::class)
