+++ app/src/testDebug/kotlin/org/cru/godtools/ui/tools/VariantToolCardPaparazziTest.kt
@@ -19,8 +19,8 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
+import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.model.Language
-import org.cru.godtools.ui.BasePaparazziTest
 import org.hamcrest.CoreMatchers.equalTo
 import org.junit.Assume.assumeThat
 import org.junit.runner.RunWith
