+++ gradle/libs.versions.toml
@@ -31,6 +31,7 @@ ktlint = "1.7.1"
 ktlintGradle = "14.0.1"
 lottie = "6.7.1"
 okhttp3 = "5.3.2"
+paparazzi = "2.0.0-alpha02"
 play-app-update = "2.1.0"
 retrofit2 = "3.0.0"
 scarlet = "0.1.12"
@@ -194,6 +195,7 @@ okhttp3 = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp3" }
 okhttp3-bom = { module = "com.squareup.okhttp3:okhttp-bom", version.ref = "okhttp3" }
 okhttp3-mockwebserver = { module = "com.squareup.okhttp3:mockwebserver", version.ref = "okhttp3" }
 okio-fakefilesystem = "com.squareup.okio:okio-fakefilesystem:3.16.4"
+paparazzi = { module = "app.cash.paparazzi:paparazzi", version.ref = "paparazzi" }
 picasso = "com.squareup.picasso:picasso:2.8"
 picasso-transformations = "jp.wasabeef:picasso-transformations:2.4.0"
 play-app-update-ktx = "com.google.android.play:app-update-ktx:2.1.0"
@@ -245,4 +247,4 @@ hilt = { id = "com.google.dagger.hilt.android", version.ref = "dagger" }
 ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 ktlint = { id = "org.jlleitschuh.gradle.ktlint", version.ref = "ktlintGradle" }
 mockposable = { id = "com.jeppeman.mockposable", version = "0.15" }
-paparazzi = { id = "app.cash.paparazzi", version = "2.0.0-alpha02" }
+paparazzi = { id = "app.cash.paparazzi", version.ref = "paparazzi" }
